safari-600.7-branch/LayoutTests/svg/dynamic-updates/SVGFilterElement-svgdom-filterRes-call-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'setFilterRes' call of the SVGFilterElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS filterElement.filterResX.baseVal is 400
PASS filterElement.filterResY.baseVal is 400
PASS filterElement.filterResX.baseVal is 200
PASS filterElement.filterResY.baseVal is 200
PASS successfullyParsed is true

TEST COMPLETE
